+++ NEWS.md
@@ -5,6 +5,7 @@
 * 11 palettes have been added from the feathers package. (#116)
 * 56 palettes have been added from the MetBrewer package. (#119)
 * 7 palettes have been added from the musculusColors package. (#120)
+* 21 palettes have been added from the peRReo package. (#122)
 
 # paletteer 1.4.1